+++ app/src/main/java/com/example/moviedb/data/repository/MovieRepositoryImpl.kt
@@ -1,9 +1,8 @@
 package com.example.moviedb.data.repository
 
-import com.example.moviedb.Constants
+import com.example.moviedb.util.Constants
 import com.example.moviedb.data.api.MovieApi
 import com.example.moviedb.model.details.MovieDetails
-import com.example.moviedb.model.movie.Movie
 import com.example.moviedb.model.movie.MovieResponse
 import retrofit2.Response
 import javax.inject.Inject
@@ -26,7 +25,7 @@ class MovieRepositoryImpl @Inject constructor(
         )
     }
 
-    override suspend fun getMovieDetails(id: Int): Response<MovieDetails> {
+    override suspend fun getMovieDetails(id: Long): Response<MovieDetails> {
         return service.getMovieDetails(
             id = id,
             apiKey = Constants.API_KEY
